feat(android): split features into colorful screens
