+++ android/app/src/main/java/com/example/lockeduntil/UseCasesActivity.kt
@@ -1,0 +1,11 @@
+package com.lockeduntil.app
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class UseCasesActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_use_cases)
+    }
+}
