+++ android/app/src/main/java/com/example/lockeduntil/WhyTrustActivity.kt
@@ -1,0 +1,11 @@
+package com.lockeduntil.app
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class WhyTrustActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_why_trust)
+    }
+}
