+++ android/app/src/main/res/layout/activity_store.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:background="@color/teal">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -14,85 +15,57 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/store_title"
-            android:textStyle="bold" />
+            android:textStyle="bold"
+            android:textColor="@color/white" />
 
         <EditText
             android:id="@+id/secretInput"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="@string/hint_secret"/>
+            android:hint="@string/hint_secret"
+            android:background="@android:color/white"/>
 
         <EditText
             android:id="@+id/emailInput"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="@string/hint_email"/>
+            android:hint="@string/hint_email"
+            android:background="@android:color/white"/>
 
         <EditText
             android:id="@+id/masterPassInput"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="@string/hint_master"/>
+            android:hint="@string/hint_master"
+            android:background="@android:color/white"/>
 
         <EditText
             android:id="@+id/viewPassInput"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="@string/hint_view"/>
+            android:hint="@string/hint_view"
+            android:background="@android:color/white"/>
 
         <EditText
             android:id="@+id/dateInput"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="@string/hint_date"/>
+            android:hint="@string/hint_date"
+            android:background="@android:color/white"/>
 
         <Button
             android:id="@+id/storeButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/store_button"/>
+            android:text="@string/store_button"
+            android:backgroundTint="@color/green"
+            android:textColor="@color/bg_dark"/>
 
         <TextView
             android:id="@+id/storeResult"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content"/>
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:background="#CCCCCC"
-            android:layout_marginTop="16dp"
-            android:layout_marginBottom="16dp"/>
-
-        <TextView
-            android:id="@+id/retrieveTitle"
-            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/retrieve_title"
-            android:textStyle="bold" />
-
-        <EditText
-            android:id="@+id/idInput"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/hint_id"/>
-
-        <EditText
-            android:id="@+id/passInput"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="@string/hint_password"/>
-
-        <Button
-            android:id="@+id/retrieveButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/retrieve_button"/>
-
-        <TextView
-            android:id="@+id/retrieveResult"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"/>
+            android:textColor="@color/white"/>
 
     </LinearLayout>
 </ScrollView>
