+++ android/app/src/main/res/values/colors.xml
@@ -1,0 +1,7 @@
+<resources>
+    <color name="bg_dark">#0f172a</color>
+    <color name="teal">#0ea5a9</color>
+    <color name="green">#10b981</color>
+    <color name="orange">#f97316</color>
+    <color name="white">#ffffff</color>
+</resources>
